deployement
Former-commit-id: 670b50d85477fa87159c7ada0667e85e536d58b1 [formerly f0f74277c76348a1c623d4722368471c989e0db8]
Former-commit-id: 681898d0fb7b15f9a8b817f55a8920720251e048
Former-commit-id: 3f438deb3db4b9586451c0f3e8a464feafae42b2 [formerly d4711697a6cc94d65106bfb9a132809ed24d380a]
Former-commit-id: 05884d86dfbf54d563049b4bac4032ae4b137e59

+++ trunk/sfark/sfarkextractor2.cpp
@@ -204,7 +204,7 @@ void OpenInputFile(const char *FileName)
 
 int ReadInputFile(quint8 *Buf, int BytesToRead)
 {
-    int	BytesRead;
+    unsigned long BytesRead;
 
     //BytesRead = SfArkExtractor2::_fileManager.read(InputFileHandle, (char*)Buf, BytesToRead);
     READFILE(InputFileHandle, Buf, BytesToRead, &BytesRead);
